+++ resale/src/main/java/com/goti/service/application/ResalePriceService.java
@@ -1,4 +1,4 @@
-package com.goti.service;
+package com.goti.service.application;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
