+++ resale/src/main/java/com/goti/service/application/ResaleTransactionService.java
@@ -1,4 +1,4 @@
-package com.goti.service;
+package com.goti.service.application;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -24,6 +24,7 @@ import com.goti.repository.ResaleRestrictionRepository;
 import com.goti.repository.ResaleTransactionRepository;
 import com.goti.repository.history.ResalePriceHistoryRepository;
 import com.goti.repository.listing.ResaleListingRepository;
+import com.goti.service.PaymentService;
 import com.goti.utils.ResalePricePolicy;
 import com.goti.utils.ResaleRestrictionHandler;
